Refactoring bot.js
closes #15
