+++ msg_handlers/README.md
@@ -1,0 +1,5 @@
+## Message handlers
+
+Each file in this directory handles bot messages, commands, etc.
+
+To add more, just create a `.js` file. **Note**: It should export a `init()` function, that will be called once automatically, where you can add all the handles to the bot (using `bot.js`).
