+++ utils.js
@@ -1,0 +1,5 @@
+let utils = module.exports = {};
+
+/* ID of group that controls this bot
+ * This could be in a environment variable... oh well */
+utils.admin_id = -212740289;
